data-pipelines/caption_text/pyproject.toml
[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[project]
name = "caption_text"
# Version is automatically determined from git tags via hatch-vcs
dynamic = ["version"]
description = "VLM-based caption text extraction and correction pipeline"
readme = "README.md"
dependencies = [
    "typer>=0.9.0",
    "rich>=13.0.0",
    "pillow>=10.0.0",
    "numpy>=1.26.0",
    "torch>=2.1.0",
    "transformers>=4.40.0",
    "peft>=0.10.0",
    "lightning>=2.0.0",
    "caption_models",
    "frames_db",
    "image_utils",
    "video_utils",
    "tqdm>=4.66.0",
]

[project.optional-dependencies]
vetting = [
    "anthropic>=0.20.0",  # For Claude API text vetting
    "ollama>=0.1.0",      # For local Ollama text vetting
]

[project.scripts]
caption_text = "caption_text.cli:app"

[tool.hatch.version]
source = "vcs"

[tool.hatch.version.raw-options]
root = "../.."

[tool.hatch.build.targets.wheel]
packages = ["src/caption_text"]

[tool.hatch.build.hooks.vcs]
version-file = "src/caption_text/_version.py"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "unit: Unit tests for individual components",
    "integration: Integration tests for multiple components",
    "slow: Slow tests requiring model loading (skipped by default)",
]

[tool.black]
line-length = 120
target-version = ["py314"]

[tool.ruff]
line-length = 120
target-version = "py314"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "B008",  # Function calls in argument defaults (typer idiom)
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["E741", "E501"]  # Allow short var names and long lines in tests
"*/_version.py" = ["UP", "I001"]  # Auto-generated file by hatch-vcs

[tool.ruff.format]
quote-style = "double"

[tool.pyright]
pythonVersion = "3.14"
typeCheckingMode = "strict"
reportUnusedImport = true
reportUnusedVariable = true
reportDuplicateImport = true

[tool.coverage.run]
source = ["src/caption_text"]
omit = ["*/tests/*"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]

[tool.uv.sources]
frames_db = { workspace = true }
image_utils = { workspace = true }
video_utils = { workspace = true }
ocr_utils = { workspace = true }
caption_models = { workspace = true }
